Refactors per PR feedback from @scottf
