+++ src/main/java/io/nats/client/support/WebsocketInputStream.java
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package io.nats.client.impl;
+package io.nats.client.support;
 
 import java.io.IOException;
 import java.io.InputStream;
