+++ src/test/java/io/nats/client/support/WebsocketTests.java
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package io.nats.client.impl;
+package io.nats.client.support;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -26,7 +26,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static io.nats.client.impl.WebsocketFrameHeader.OpCode;
+import static io.nats.client.support.WebsocketFrameHeader.OpCode;
 
 public class WebsocketTests {
     private int[] testSizes = new int[] {
